.requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
oauthlib==3.2.2
pip==24.2
psycopg-binary==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
wheel==0.44.0